overlay: use overlay check fs helpers in tests
Replace explicit checks in tests that work and index dirs are empty
with check fs helpers that check that both dirs are 'clean'.

With upcoming NFS export support, after unlinking all overlay
handlinks, instead of removal of index entry, it is replaced with
a whiteout index entry.

The helpers are compatible with this upcoming change because they
ignore chardev index entries when checking for 'clean' index dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>