overlay/036: factor out test_exclusive_dir() helper
In test overlay/036 there is a sub-test for exclusive upperdir/workdir
that is repeated for index=off and index=on. The sub-test uses a "first"
mount using device name "overlay" and tries to execute two other mounts
that reuse upperdir/workdir respectively.

Create a helper to execute the sub-test and call the helper with the two
mount option variations.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>