overlay: add a test for multiple redirects to the same lower dir
Multiple redirects to the same lower dir will falsely return the
same st_ino/st_dev for two different upper dirs and will cause 'diff'
to falsely report that content of directories is the same when it is not.

This test checks that overalyfs detects and fails lookup of a multiply
redirected dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>